+++ frontend/app/src/components/reranker/CreateRerankerForm.tsx
@@ -4,7 +4,7 @@ import { type CreateReranker, createReranker, listRerankerOptions, type Reranker
 import { ProviderSelect } from '@/components/form/biz';
 import { FormInput } from '@/components/form/control-widget';
 import { formFieldLayout } from '@/components/form/field-layout';
-import { FormRootErrorBeta as FormRootError } from '@/components/form/root-error';
+import { FormRootError } from '@/components/form/root-error';
 import { onSubmitHelper } from '@/components/form/utils';
 import { CodeInput } from '@/components/form/widgets/CodeInput';
 import { ProviderDescription } from '@/components/provider-description';
@@ -12,7 +12,7 @@ import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from '@/
 import { Form, formDomEventHandlers, FormSubmit } from '@/components/ui/form.beta';
 import { useModelProvider } from '@/hooks/use-model-provider';
 import { zodJsonText } from '@/lib/zod';
-import { useForm as useTanstackForm } from '@tanstack/react-form';
+import { useForm } from '@tanstack/react-form';
 import { useId, useState } from 'react';
 import { toast } from 'sonner';
 import useSWR from 'swr';
@@ -43,7 +43,7 @@ export function CreateRerankerForm ({ transitioning, onCreated }: { transitionin
   const { data: options, isLoading, error } = useSWR('api.rerankers.list-options', listRerankerOptions);
   const [submissionError, setSubmissionError] = useState<unknown>();
 
-  const form = useTanstackForm<CreateReranker | Omit<CreateReranker, 'model' | 'credentials'>>({
+  const form = useForm<CreateReranker | Omit<CreateReranker, 'model' | 'credentials'>>({
     validators: {
       onSubmit: unsetForm,
     },
