+++ frontend/app/src/components/settings-form/index.ts
@@ -1,4 +1,4 @@
 export { GeneralSettingsForm } from './GeneralSettingsForm';
-export { fieldAccessor, GeneralSettingsFieldBeta as GeneralSettingsField, GeneralSettingsFieldBeta, type GeneralSettingsFieldAccessor } from './GeneralSettingsField';
+export { fieldAccessor, GeneralSettingsField, type GeneralSettingsFieldAccessor } from './GeneralSettingsField';
 export { shallowPick } from './utils';
 export { createAccessorHelper } from './accessor-helper';
